--- Assets/Scenes/Tareas/Cables.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d9cb8fbb52f53e8438464c437c8e4383
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
