+++ Assets/Scenes/Tareas/Cables/CableFinal.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4c1448a1bd0339e1cb96d6e27d9f925f57dcaf19131acc5b791fa57a8edb0726
+size 24695
