--- Assets/Scenes/Tareas/Cables/Intento1.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: afb51969427f2d8409b925b0ea815f64
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
