--- Assets/Scenes/Tareas/Cables/Place Holders Cables.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 771b39e47c1b695439c4a8df4748e106
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
