+++ Assets/Scenes/Tareas/Cables/RandomizarCables.cs.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 9d3e8d3a5e2534646b45a553dbfd037c
+guid: 9eff140b25d64601aabc6ba32245d099
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
