--- Assets/Scenes/Tareas/Place Holders Cables.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 57f951aac874c4e4bb37dc202e5ebb08
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
