--- Assets/Scenes/Tareas/Place Holders Cables/Cable.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ca781f2a0ed1ceb8edb8cc94c32e30c2574f7282b3fef0a6724360a347fda8e5
-size 8778
