--- Assets/Scenes/Tareas/Place Holders Cables/Cable.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8f7361b89f5850341b508ef6ab9387af
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
