--- Assets/Scenes/Tareas/Place Holders Cables/Caja.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:973572e59e631b0cd7564e4f1995820eab357391f45d665355d17bc80adbaae5
-size 5281
