+++ Assets/TextMesh Pro/Examples _ Extras/Scripts/SimpleScript.cs.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 9eff140b25d64601aabc6ba32245d099
+guid: 39ec07c5e3659c048bfa796e1575a3e2
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
